rework components and app startup to include routing, initialize firebase config
