+++ package.json
@@ -11,6 +11,9 @@
   "dependencies": {
     "@expo/config-plugins": "^7.2.5",
     "@expo/webpack-config": "^19.0.0",
+    "@react-native-async-storage/async-storage": "1.18.2",
+    "@react-navigation/native": "^6.1.9",
+    "@react-navigation/native-stack": "^6.9.17",
     "@tensorflow-models/mobilenet": "^2.1.1",
     "@tensorflow/tfjs": "^4.15.0",
     "@tensorflow/tfjs-react-native": "^1.0.0",
@@ -25,7 +28,6 @@
     "expo-image-manipulator": "~11.3.0",
     "expo-image-picker": "~14.3.2",
     "expo-media-library": "~15.4.1",
-    "expo-status-bar": "~1.6.0",
     "firebase": "^10.7.1",
     "openai": "^4.24.3",
     "react": "18.2.0",
@@ -33,6 +35,7 @@
     "react-native": "^0.72.6",
     "react-native-fs": "^2.20.0",
     "react-native-safe-area-context": "4.6.3",
+    "react-native-screens": "~3.22.0",
     "react-native-url-polyfill": "^2.0.0",
     "react-native-web": "~0.19.6"
   },
